adds openai dataset functionality
